jenkinsreport/requirements.txt
#INSTALL PIP PACKAGES
celery==4.2.0
redis==3.2.1
Django==1.8
jira==2.0.0
django-celery-beat==1.1.1
python-jenkins==1.4.0

#INSTALL redis-server
#sudo apt-get install redis-server

